Fixes #3533 Consume OSPSuite.Core ApplicationIcon refactor
Bumps OSPSuite.* references to 13.0.110, which ships both #2858
(ApplicationIcon is a pure cross-platform resource library, Windows rendering
relocated) and #2864 (ApplicationIconExtensions moved from OSPSuite.UI to
OSPSuite.Presentation).

Mechanical call-site updates for the removed implicit ApplicationIcon →
SvgImage / Image / Size operators: explicit .ToSvgImage() / .ToImage() /
.ToDrawingSize() / .SetImage(...) throughout PKSim.UI.

DTOs that expose System.Drawing.Image properties bound into DevExpress
PictureEdit grid columns (SimulationSystemicProcessSelectionDTO,
SimulationPartialProcessSelectionDTO, PopulationFileSelectionDTO,
SimulationResultsFileSelectionDTO) now call `icon.ToImage()` from
OSPSuite.Presentation.Extensions directly. Drops the temporary
PKSim.Presentation/DTO/IconImageRenderer.cs helper and the explicit
DevExpress.Utils PackageReference on PKSim.Presentation (it inherits
transitively from OSPSuite.Presentation now).

TFM bumps to net10.0-windows where forced by transitive OSPSuite.Presentation
deps (PKSim.Presentation, PKSim.Infrastructure, PKSim.Assets.Images).
PKSim.R / PKSim.R.Tests bumped temporarily with explanatory comments
pointing at #3537 (R cross-platform needs OSPSuite.Presentation pull removed
*and* PKSim.Core System.Drawing surface refactored — PKSimColors,
ColorGenerator, chart/snapshot models — see issue comment for the full audit).

PKSim.Assets.Images project eliminated: ApplicationImages (the raster-PNG
helper) and its two embedded PNGs move under PKSim.UI/Views/Simulations/;
installer icons (PKSim.ico, JournalFile.ico, PKSim5FileIcon.ico, *.svg) move
to PKSim.UI/Resources/ and rakefile setup glob updated to match;
About.bmp deleted as dead. The deleted ApplicationImage wrapper class from
OSPSuite.Assets is replaced by direct use of System.Drawing.Image inside the
relocated ApplicationImages.cs.
