+++ PKSim.sln
@@ -47,8 +47,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PKSim.CLI.Core", "src\PKSim
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PKSim.Tests", "tests\PKSim.Tests\PKSim.Tests.csproj", "{590B08C3-CAF6-4A90-9BFE-60EEBEC3C661}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PKSim.Assets.Images", "src\PKSim.Assets.Images\PKSim.Assets.Images.csproj", "{95E92AA4-E2A3-4E60-BD5A-E0742937EA05}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PKSim.R", "src\PKSim.R\PKSim.R.csproj", "{AEE05F2D-C709-4F03-BFDE-14FF68744F35}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PKSim.R.Tests", "tests\PKSim.R.Tests\PKSim.R.Tests.csproj", "{4E310F99-4604-4502-8F0B-A4E4F27B062B}"
@@ -105,10 +103,6 @@ Global
 		{590B08C3-CAF6-4A90-9BFE-60EEBEC3C661}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{590B08C3-CAF6-4A90-9BFE-60EEBEC3C661}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{590B08C3-CAF6-4A90-9BFE-60EEBEC3C661}.Release|Any CPU.Build.0 = Release|Any CPU
-		{95E92AA4-E2A3-4E60-BD5A-E0742937EA05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{95E92AA4-E2A3-4E60-BD5A-E0742937EA05}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{95E92AA4-E2A3-4E60-BD5A-E0742937EA05}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{95E92AA4-E2A3-4E60-BD5A-E0742937EA05}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AEE05F2D-C709-4F03-BFDE-14FF68744F35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AEE05F2D-C709-4F03-BFDE-14FF68744F35}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AEE05F2D-C709-4F03-BFDE-14FF68744F35}.Release|Any CPU.ActiveCfg = Release|Any CPU
