--- src/PKSim.Assets.Images/PKSim.Assets.Images.csproj
@@ -1,64 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
-    <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
-    <GenerateResourceUsePreserializedResources>true</GenerateResourceUsePreserializedResources>
-    <PackageTags>open-systems-pharmacology, ospsuite-components</PackageTags>
-    <Authors>Open-Systems-Pharmacology</Authors>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <OutputPath>bin\$(Configuration)</OutputPath>
-    <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\PKSim.Core.xml</DocumentationFile>
-    <NoWarn>1591</NoWarn>
-    <Version Condition="'$(Version)' == ''">13.0.0</Version>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Resources\SmallMoleculesStructure.png" />
-    <None Remove="Resources\TwoPoresModelStructure.png" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="..\..\SolutionInfo.cs" Link="Properties\SolutionInfo.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Properties\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Assets.Images" Version="13.0.108" />
-    <PackageReference Include="System.Resources.Extensions" Version="9.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\..\LICENSE">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Resources\SmallMoleculesStructure.png" />
-    <EmbeddedResource Include="Resources\TwoPoresModelStructure.png" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Images.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Images.resx</DependentUpon>
-    </Compile>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Update="Images.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Images.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-  </ItemGroup>
-
-
-</Project>
