--- src/PKSim.Assets.Images/PKSim.Assets.Images.v3.ncrunchproject
@@ -1,5 +1,0 @@
-﻿<ProjectConfiguration>
-  <Settings>
-    <PreviouslyBuiltSuccessfully>True</PreviouslyBuiltSuccessfully>
-  </Settings>
-</ProjectConfiguration>
