+++ src/PKSim.Core/PKSim.Core.csproj
@@ -26,10 +26,10 @@
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="OSPSuite.Utility" Version="5.0.0.7" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Assets.Images" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure.Import" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Assets.Images" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure.Import" Version="13.0.110" />
     <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
   </ItemGroup>
