+++ src/PKSim.Infrastructure/PKSim.Infrastructure.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
+    <TargetFramework>net10.0-windows</TargetFramework>
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
     <PackageTags>open-systems-pharmacology, ospsuite-components</PackageTags>
@@ -36,14 +36,14 @@
     <PackageReference Include="Newtonsoft.Json.Schema" Version="4.0.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="NHibernate.Extensions.Sqlite" Version="10.0.1" />
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure.Export" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure.Import" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure.Serialization" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Presentation.Serialization" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure.Export" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure.Import" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure.Serialization" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Presentation.Serialization" Version="13.0.110" />
     <PackageReference Include="OSPSuite.Utility" Version="5.0.0.7" />
     <PackageReference Include="Microsoft.Data.Sqlite" Version="10.0.7" />
   </ItemGroup>
