+++ src/PKSim.R/PKSim.R.csproj
@@ -1,7 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
+    <!--
+      Temporary: PKSim.R is meant to be cross-platform net10.0, but transitively pulls
+      OSPSuite.Presentation (Windows-only) via PKSim.Infrastructure, and also pulls a
+      System.Drawing surface from PKSim.Core (PKSimColors, ColorGenerator, chart/snapshot
+      models). Once both are addressed, this TFM returns to net10.0 and the assembly
+      loads honestly on Linux / macOS .NET 10.
+      Tracked in https://github.com/Open-Systems-Pharmacology/PK-Sim/issues/3537.
+    -->
+    <TargetFramework>net10.0-windows</TargetFramework>
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
     <PackageTags>open-systems-pharmacology, ospsuite-components</PackageTags>
@@ -36,10 +44,10 @@
 
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.R" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.R" Version="13.0.110" />
     <PackageReference Include="OSPSuite.Utility" Version="5.0.0.7" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.110" />
     <PackageReference Include="OSPSuite.FuncParser" Version="5.0.0.3" />
     <PackageReference Include="OSPSuite.SimModel" Version="5.0.0.76" />
     <PackageReference Include="OSPSuite.SimModelSolver_CVODES" Version="5.0.0.53" />
