+++ src/PKSim.Starter/PKSim.Starter.csproj
@@ -20,10 +20,10 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Presentation" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.UI" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Presentation" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.UI" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.110" />
   </ItemGroup>
 
 </Project>
