+++ src/PKSim.UI/PKSim.UI.csproj
@@ -46,13 +46,13 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.2" />
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.2" />
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.110" />
     <PackageReference Include="OSPSuite.DataBinding" Version="4.0.0.5" />
     <PackageReference Include="OSPSuite.DataBinding.DevExpress" Version="7.0.0.6" />
     <PackageReference Include="OSPSuite.Utility" Version="5.0.0.7" />
-    <PackageReference Include="OSPSuite.Presentation" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.UI" Version="13.0.108" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.108" />
+    <PackageReference Include="OSPSuite.Presentation" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.UI" Version="13.0.110" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.110" />
     <PackageReference Include="DevExpress.Win.Design" Version="25.2.7" Condition="'$(ExcludeDesigner)' != 'true'" />
   </ItemGroup>
   <ItemGroup>
@@ -67,9 +67,13 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\PKSim.Assets.Images\PKSim.Assets.Images.csproj" />
     <ProjectReference Include="..\PKSim.Core\PKSim.Core.csproj" />
     <ProjectReference Include="..\PKSim.Infrastructure\PKSim.Infrastructure.csproj" />
     <ProjectReference Include="..\PKSim.Presentation\PKSim.Presentation.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <!-- Model-structure PNGs consumed by SimulationModelSelectionView via ApplicationImages. -->
+    <EmbeddedResource Include="Views\Simulations\Resources\SmallMoleculesStructure.png" />
+    <EmbeddedResource Include="Views\Simulations\Resources\TwoPoresModelStructure.png" />
   </ItemGroup>
 </Project>
