+++ src/PKSim.UI/Resources/PKSim5FileIcon.ico
