+++ src/PKSim.UI/Views/Core/UxDropDownButton.cs
@@ -5,6 +5,8 @@ using DevExpress.Utils;
 using DevExpress.Utils.Menu;
 using DevExpress.XtraEditors;
 using OSPSuite.Assets;
+using OSPSuite.Presentation.Extensions;
+using OSPSuite.UI.Extensions;
 using OSPSuite.Utility.Container;
 using OSPSuite.Utility.Exceptions;
 using OSPSuite.Utility.Extensions;
