+++ src/PKSim.UI/Views/Simulations/Resources/TwoPoresModelStructure.png
