+++ src/PKSim.UI/Views/Simulations/SimulationModelSelectionView.cs
@@ -5,7 +5,6 @@ using PKSim.Presentation.DTO.Simulations;
 using PKSim.Presentation.Presenters.Simulations;
 using PKSim.Presentation.Views.Simulations;
 using OSPSuite.UI.Controls;
-using PKSim.Assets.Images;
 
 namespace PKSim.UI.Views.Simulations
 {
