+++ tests/PKSim.UI.Tests/PKSim.UI.Tests.csproj
@@ -22,11 +22,11 @@
 		</PackageReference>
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
 		<PackageReference Include="nunit" Version="3.14.0" />
-		<PackageReference Include="OSPSuite.Assets" Version="13.0.108" />
+		<PackageReference Include="OSPSuite.Assets" Version="13.0.110" />
 		<PackageReference Include="OSPSuite.BDDHelper" Version="4.0.1.1" />
-		<PackageReference Include="OSPSuite.Core" Version="13.0.108" />
+		<PackageReference Include="OSPSuite.Core" Version="13.0.110" />
 		<PackageReference Include="OSPSuite.SimModel" Version="5.0.0.76" />
-		<PackageReference Include="OSPSuite.UI" Version="13.0.108" />
+		<PackageReference Include="OSPSuite.UI" Version="13.0.110" />
 		<PackageReference Include="OSPSuite.FuncParser" Version="5.0.0.3" />
 		<PackageReference Include="Microsoft.Data.Sqlite" Version="10.0.7" />
 		<None Include="..\..\src\Db\PKSimDB.sqlite" Link="PKSimDB.sqlite">
